requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
scapy>=2.5.0
pillow>=10.0.0
cryptography>=41.0.0
pycryptodome>=3.19.0
colorama>=0.4.6
termcolor>=2.3.0
tqdm>=4.66.0
python-nmap>=0.7.1
paramiko>=3.3.0
pyshark>=0.6
netifaces>=0.11.0
psutil>=5.9.0
watchdog>=3.0.0
python-magic>=0.4.27
dnspython>=2.4.0
shodan>=1.30.0
censys>=2.2.0
pefile>=2023.2.7
yara-python>=4.3.1
numpy>=1.24.0
stem>=1.8.0
pysocks>=1.7.1


